+++ nata-monkey/src/main/java/com/nata/monkeys/RandomMonkey.java
@@ -43,16 +43,16 @@ public class RandomMonkey extends AbstractMonkey {
         System.out.println("start playing...");
         startApp();
         while(isRunning){
-//            Action nextAction = getNextUIAction();
-//            if(nextAction==  null){
-//                nextAction = new BackAction(getDevice());
-//            }
-            Action nextAction ;
-            if(Math.random() > 0.5){
-                nextAction = new SwipeAction(SwipeDirection.RIGHT,getDevice());
-            }else{
-                nextAction = new SwipeAction(SwipeDirection.LEFT,getDevice());
+            Action nextAction = getNextUIAction();
+            if(nextAction==  null){
+                nextAction = new BackAction(getDevice());
             }
+//            Action nextAction ;
+//            if(Math.random() > 0.5){
+//                nextAction = new SwipeAction(SwipeDirection.RIGHT,getDevice());
+//            }else{
+//                nextAction = new SwipeAction(SwipeDirection.LEFT,getDevice());
+//            }
             actionList.add(nextAction);
             nextAction.fire();
             System.out.println(nextAction);
@@ -72,7 +72,7 @@ public class RandomMonkey extends AbstractMonkey {
                 if(node.getClassName().equals("android.widget.Button") && node.getClickable().equals("true")){
                     actionList.add(new UINodeAction(node, ActionType.TAP));
                 }
-                if(node.getClassName().equals("android.widget.EditText")){
+                if(node.getClassName().equals("android.widget.EditText") && node.getText().length()==0){
                     actionList.add(new UINodeAction(node, ActionType.INPUT));
                 }
             }