+++ CHANGELOG.md
@@ -22,3 +22,4 @@
 ### v0.0.6(2016/3/8)
 - Add Swipe Action(UP,DOWN,LEFT,RIGHT)
 - Can get Device Resolution
+- Add Input Dictionary including valid and invalid inputs